removing erroneous types & fixing timezone select 🌄
